+++ src/io/driver.rs
@@ -24,6 +24,9 @@ use std::{
 /// Default timeout in seconds for read/write operations
 const DEFAULT_TIMEOUT_SECONDS: u64 = 5;
 
+/// Printer driver trait
+///
+/// A custom driver can be implemented by implementing this trait.
 pub trait Driver {
     /// Driver name
     fn name(&self) -> String;