+++ examples/ui.rs
@@ -1,6 +1,6 @@
 use escpos::printer::Printer;
 use escpos::printer_options::PrinterOptions;
-use escpos::utils::ui::line::{LineBuilder, LineStyle};
+use escpos::ui::line::{LineBuilder, LineStyle};
 use escpos::utils::*;
 use escpos::{driver::*, errors::Result};
 
@@ -22,10 +22,9 @@ fn main() -> Result<()> {
         .width(8)
         .build();
     let line_custom = LineBuilder::new().style(LineStyle::Custom("=-")).build();
-    // let line_custom_utf8 = LineBuilder::new().style(LineStyle::Custom("😀")).width(8).build();
 
-    let driver = NetworkDriver::open("192.168.1.248", 9100, None)?;
-    // let driver = ConsoleDriver::open(true);
+    // let driver = NetworkDriver::open("192.168.1.248", 9100, None)?;
+    let driver = ConsoleDriver::open(true);
     let printer_options = PrinterOptions::new(Some(PageCode::PC858), Some(DebugMode::Dec), 42);
     let mut printer = Printer::new(driver, Protocol::default(), Some(printer_options));
     printer
@@ -39,7 +38,6 @@ fn main() -> Result<()> {
         .draw_line(line_dashed)?
         .draw_line(line_custom)?
         .draw_line(line_dotted)?
-        // .draw_line(line_custom_utf8)?
         .print_cut()?;
 
     Ok(())