+++ examples/tauri-app/src-tauri/src/lib.rs
@@ -1,36 +1,14 @@
-// Learn more about Tauri commands at https://tauri.app/develop/calling-rust/
-
 mod printer;
 
-use escpos::driver::UsbDriver;
-use escpos::errors::PrinterError;
-use escpos::printer::Printer;
-use escpos::utils::Protocol;
+use crate::printer::{UsbPrinter, PRINTER_PID, PRINTER_VID};
 use printer::{print_test, printer_status};
 use std::sync::{Arc, Mutex};
-use std::time::Duration;
 use tauri::Manager;
-
-const VID: u16 = 0x0416;
-const PID: u16 = 0x5011;
-
-pub struct MyPrinter {
-    port: Printer<UsbDriver>,
-}
-
-impl MyPrinter {
-    pub fn build(vid: u16, pid: u16) -> Result<Self, PrinterError> {
-        let driver = UsbDriver::open(vid, pid, Some(Duration::from_secs(2)))?;
-        let printer = Printer::new(driver, Protocol::default(), None);
-
-        Ok(Self { port: printer })
-    }
-}
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     let state = Arc::new(Mutex::new(
-        MyPrinter::build(VID, PID).expect("error while building printer"),
+        UsbPrinter::build(PRINTER_VID, PRINTER_PID).expect("error while building printer"),
     ));
 
     tauri::Builder::default()