+++ examples/tauri-app/src-tauri/src/printer.rs
@@ -1,4 +1,4 @@
-use escpos::driver::{Driver, NetworkDriver};
+use escpos::driver::{Driver, NetworkDriver, UsbDriver};
 use escpos::errors::PrinterError;
 use escpos::printer::Printer;
 use escpos::utils::{Protocol, RealTimeStatusRequest, RealTimeStatusResponse};
@@ -7,6 +7,21 @@ use std::time::Duration;
 
 pub const PRINTER_ADDR: &str = "192.168.1.248";
 pub const PRINTER_PORT: u16 = 9100;
+pub const PRINTER_VID: u16 = 0x0525;
+pub const PRINTER_PID: u16 = 0xa700;
+
+pub struct UsbPrinter {
+    port: Printer<UsbDriver>,
+}
+
+impl UsbPrinter {
+    pub fn build(vid: u16, pid: u16) -> Result<Self, PrinterError> {
+        let driver = UsbDriver::open(vid, pid, Some(Duration::from_secs(2)))?;
+        let printer = Printer::new(driver, Protocol::default(), None);
+
+        Ok(Self { port: printer })
+    }
+}
 
 #[derive(Debug, Serialize)]
 pub struct CustomError {
@@ -23,8 +38,9 @@ impl From<PrinterError> for CustomError {
 
 #[tauri::command]
 pub async fn print_test() -> Result<(), CustomError> {
-    let driver = NetworkDriver::open(PRINTER_ADDR, PRINTER_PORT, Some(Duration::from_secs(2)))?;
-    let mut printer = Printer::new(driver, Protocol::default(), None);
+    let driver = NetworkDriver::open(PRINTER_ADDR, PRINTER_PORT, Some(Duration::from_secs(1)))?;
+    let mut printer = Printer::new(driver.clone(), Protocol::default(), None);
+
     printer.init()?.writeln("test")?.feed()?.print_cut()?;
 
     Ok(())