+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Changed (for changes in existing functionality)
+### Changed
 
 - [BREAKING] Add `PrinterOptions` to `Printer` instead of `PageCode`  
   Before:
@@ -36,12 +36,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ```rust
   let mut printer = Printer::new(driver, Protocol::default(), None);
   ```
-- Remove `lazy_static` and use standard library
+- Remove `lazy_static` and use standard library `LazyLock` instead
 - Bump `image` to `0.25.2`
 - Bump `nusb` to `0.1.10`
 - Bump `env_logger` to `0.11.5`
 - Bump `hidapi` to `2.6.3`
 - Bump `serialport` to `1.5.0`
+
+### Fixed
+
+- Fix documentation
+-
 
 ## `0.12.2` (2024-04-23) [CURRENT]