+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add a new UI `Line` component
 - Add a state for printer styles
 
+### Changed
+
+- Update examples
+
 ## `0.13.0` (2024-08-08) [CURRENT]
 
 ### Changed